requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
httpx==0.28.1

# OpenAI
openai==1.6.1

# Vector store
faiss-cpu==1.7.4

# NLP
spacy==3.7.2

# Optional: NLI scoring (transformers)
# transformers==4.36.2
# torch==2.1.2

# Streamlit UI
streamlit==1.29.0

